duplicate command fixed, ctr+d doesnt work anymore, ctrl+\ core dumped client side
